♻️ Refactor : calculateDday 함수 분리
