+++ src/utils/formatDate.ts
@@ -27,3 +27,20 @@ export const formatDatePost = (dateTime: string) => {
     return formatDateToYYYYMMDD(dateTime)
   }
 }
+
+export const calculateDday = (deadline: string) => {
+  const today = new Date()
+  today.setHours(0, 0, 0, 0)
+  const target = new Date(deadline)
+  target.setHours(0, 0, 0, 0)
+  const day = 24 * 60 * 60 * 1000
+  const diffDays = Math.ceil((target.getTime() - today.getTime()) / day)
+
+  if (diffDays === 0) {
+    return 'D-day'
+  } else if (diffDays > 0) {
+    return `D-${diffDays}`
+  } else {
+    return '모집이 완료되었습니다.'
+  }
+}
